requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.30
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-dotenv==1.0.1
deepface==0.0.93
mediapipe>=0.10.30
numpy>=2.0.0
opencv-python-headless>=4.10.0
cloudinary==1.40.0
firebase-admin==6.5.0
httpx==0.27.0
pillow>=10.4.0
tf-keras>=2.16.0
tensorflow>=2.16.1
pgvector==0.2.5
